Fixed data type in interface

+++ src/main/java/de/fernunihagen/dna/jkn/scalephant/storage/sstable/compact/MergeStrategy.java
@@ -16,6 +16,6 @@ public interface MergeStrategy {
 	 * Get the delay for the compact thread
 	 * @return
 	 */
-	public abstract int getCompactorDelay();
+	public abstract long getCompactorDelay();
 
 }
